+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # jekyll-humanize Changelog
 
+## v1.1.0 (2014-07-14)
+
++ Added new filesize filter for converting bytes
++ Resolved #1 to add `<sup></sup>` support to ordinal numbers
+- Small updates to README.md
+
 ## v1.0.1 (2014-01-29)
 
 + Fixed small type in gsub (proper escaping needed)